requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
fastapi==0.115.12
h11==0.16.0
idna==3.10
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
psutil==7.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pyinstaller==6.14.0
pyinstaller-hooks-contrib==2025.4
pypiwin32==223
pystray==0.19.5
pywin32==310
pywin32-ctypes==0.2.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
websockets==15.0.1
win10toast==0.9
